cleanup for pull request
