+++ src/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "build": "tsc",
     "server": "node ./dist/app.js",
-    "dev": "tsc --watch & nodemon ./dist/app.js"
+    "dev": "tsc --watch & nodemon dist/app.js"
   },
   "dependencies": {
     "@fastify/cors": "^11.0.1",
@@ -19,6 +19,7 @@
     "better-sqlite3": "^11.9.1",
     "ejs": "^3.1.10",
     "fastify": "^5.2.2",
+    "nodemon": "^3.1.9",
     "typescript": "^5.8.3"
   }
 }
