+++ src/views/index.html
@@ -2,19 +2,26 @@
 <html lang="en">
 
 <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>ft_transcendence</title>
-    <link rel="icon" href="/public/favicon.ico">
+	<meta charset="UTF-8" />
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<title>Pong Game</title>
+	<link rel="icon" href="/public/favicon.ico">
     <link rel="stylesheet" href="/public/styles.css">
+	<style>
+		body { background: black; margin: 0; overflow: hidden; }
+		canvas { display: block; margin: auto; background: #111; }
+	</style>
 </head>
 
 <body>
-    <h1>Tic Tac Toe with Typescript</h1>
-    <div class="gameBoard"></div>
-    <button class="restartButton">Restart Game</button>
-    <!-- <script type="module" src="/dist/views/main.js"></script> -->
-    <script type="module" src="/dist/views/main_api.js"></script>
+	<h1>Pong</h1>
+	<button id="startGameBtn">Start Game</button>
+	<button id="pauseGameBtn">Pause Game</button>
+	<button id="resumeGameBtn">Resume Game</button>
+	<button id="resetGameBtn">Reset Game</button>
+	<canvas id="gameCanvas" width="800" height="600">
+		Your browser does not support the HTML5 Canvas element.
+	</canvas>
+	<script type="module" src="/dist/views/main.js"></script>
 </body>
-
 </html>
