+++ src/views/main.ts
@@ -1,74 +1,130 @@
 // const socket: WebSocket = new WebSocket("ws://localhost:3000/ws");
-const socket: WebSocket = new WebSocket("ws://10.11.2.31:3000/ws"); // 1-B-31
-
-const headerTitle = document.querySelector("h1") as HTMLHeadingElement;
-const gameBoard = document.querySelector(".gameBoard") as HTMLDivElement;
-const restartButton = document.querySelector(".restartButton") as HTMLButtonElement;
-let cells: HTMLDivElement[] = [];
-
-// The "open" event is triggered when the connection to the WebSocket server is successfully established.
-socket.addEventListener("open", () => {
-    console.log("Connected to WebSocket server");
+const socket: WebSocket = new WebSocket("ws://10.11.2.27:3000/ws");
+const canvas = document.getElementById("gameCanvas") as HTMLCanvasElement;
+const ctx = canvas.getContext("2d")!;
+let state: any = null;
+let playerId: number | null = null;
+let keysPressed: Record<string, boolean> = {};
+import { PADDLE_WIDTH, PADDLE_HEIGHT, BALL_RADIUS } from "../models/Constants.js";
+//listen for websocket connection
+socket.addEventListener("open", ()=> {
+	console.log("connected to pong server");
 });
 
-// The "message" event is triggered when the server sends a message over WebSocket.
 socket.addEventListener("message", (event: MessageEvent<string>) => {
-    const data = JSON.parse(event.data);
+	const data = JSON.parse(event.data);
 
-    if (data.type === "initBoard" || data.type === "resetBoard") {
-        updateBoard(data.board);
-        updateStatus("Tic Tac Toe with Typescript");
-    }
-    else if (data.type === "updateBoard") {
-        updateBoard(data.board);
+	if (data.type === "assignPlayer") {
+		playerId = data.id;
+		console.log("Assigned as Player", playerId);
+	}
 
-        if (data.win) {
-            updateStatus(`${data.player} wins!`);
-        }
-        else if (data.draw) {
-            updateStatus("It's a draw!");
-        }
-        else {
-            updateStatus(`Current player: ${data.player}`);
-        }
-    }
+	if (data.type === "initGame" || data.type === "update") {
+		state = data.state;
+		draw();
+		console.log("Game started.");
+	}
+
+	if (data.type === "pauseGame") {
+		state = data.state;
+		draw();
+		console.log("Game paused.");
+	}
+
+	if (data.type === "resumeGame") {
+		state = data.state;
+		draw();
+		console.log("Game restarted.");
+	}
+
+	if (data.type === "resetGame") {
+		state = data.state;
+		draw();
+		console.log("Game resumed.");
+	}
 });
 
-function updateBoard(board: string[]): void {
-    cells.forEach((cell, index) => {
-        cell.textContent = board[index];
-    });
+socket.addEventListener("error", (err) => {
+	console.error("WebSocket error:", err);
+});
+
+window.addEventListener("keydown", (e) => {
+	keysPressed[e.key] = true;
+});
+
+window.addEventListener("keyup", (e) => {
+	keysPressed[e.key] = false;
+});
+
+function handleInput() {
+	if (playerId === 1) {
+		if (keysPressed["w"]) {
+			socket.send(JSON.stringify({ type: "movePaddle", direction: "up" }));
+		}
+		if (keysPressed["s"]) {
+			socket.send(JSON.stringify({ type: "movePaddle", direction: "down" }));
+		}
+	} else if (playerId === 2) {
+		if (keysPressed["ArrowUp"]) {
+			socket.send(JSON.stringify({ type: "movePaddle", direction: "up" }));
+		}
+		if (keysPressed["ArrowDown"]) {
+			socket.send(JSON.stringify({ type: "movePaddle", direction: "down" }));
+		}
+	}
 }
 
-function updateStatus(message: string): void {
-    headerTitle.textContent = message;
+setInterval(handleInput, 1000 / 60);
+
+
+function draw() {
+	if (!state || state.paused) return;
+
+	ctx.clearRect(0, 0, canvas.width, canvas.height);
+
+	// Draw ball
+	ctx.beginPath();
+	ctx.arc(state.ball.x, state.ball.y, state.ball.radius, 0, Math.PI * 2);
+	ctx.fillStyle = "#FFFFFF";
+	ctx.fill();
+	ctx.closePath();
+
+	// Draw paddles (filled)
+	ctx.fillStyle = "#FF0000";
+	ctx.fillRect(state.paddle1.x, state.paddle1.y, state.paddle1.width, state.paddle1.height);
+	ctx.fillRect(state.paddle2.x, state.paddle2.y, state.paddle2.width, state.paddle2.height);
+
+	// Add paddle hitbox outlines
+	ctx.strokeStyle = "#00FF00"; // Green outline for debugging
+	ctx.strokeRect(state.paddle1.x, state.paddle1.y, state.paddle1.width, state.paddle1.height);
+	ctx.strokeRect(state.paddle2.x, state.paddle2.y, state.paddle2.width, state.paddle2.height);
+
+	// Draw scores
+	ctx.font = "24px Arial";
+	ctx.fillStyle = "#FFFFFF";
+	ctx.fillText(`Player 1: ${state.score1}`, 50, 30);
+	ctx.fillText(`Player 2: ${state.score2}`, canvas.width - 180, 30);
 }
 
-// Event listeners for the game board.
-gameBoard.addEventListener("click", (event: MouseEvent) => {
-    const target = event.target as HTMLDivElement;
-    if (target.classList.contains("cell")) {
-        const index = Number(target.dataset.index);
-        socket.send(JSON.stringify({ type: "makeMove", index }));
-    }
-});
+  const startGameBtn = document.getElementById("startGameBtn") as HTMLButtonElement;
+  startGameBtn.addEventListener("click", () => {
+	// if (playerId) {
+		socket.send(JSON.stringify({ type: "initGame" })); // Send init message to the server
+	// }
+  });
 
-// Event listeners for the restart button.
-restartButton.addEventListener("click", () => {
-    socket.send(JSON.stringify({ type: "resetGame" }));
-});
+  const pauseGameBtn = document.getElementById("pauseGameBtn") as HTMLButtonElement;
+  pauseGameBtn.addEventListener("click", () => {
+	  socket.send(JSON.stringify({ type: "pauseGame" }));
+  });
 
-function createBoard(): void {
-    for (let i = 0; i < 9; i++) {
-        const cell: HTMLDivElement = document.createElement("div");     // Create a div element for each cell
-        cell.classList.add("cell");                                     // Add the "cell" class to style it
-        cell.dataset.index = i.toString();                              // Store the index as a string
-        gameBoard.appendChild(cell);                                    // Append the cell to the game board
-        cells.push(cell);
-    }
-}
+  const resumeGameBtn = document.getElementById("resumeGameBtn") as HTMLButtonElement;
+  resumeGameBtn.addEventListener("click", () => {
+	  socket.send(JSON.stringify({ type: "resumeGame" }));
+  });
 
-// Initialize the board when the page loads.
-window.addEventListener("DOMContentLoaded", () => {
-    createBoard();
-});
+  const resetGameBtn = document.getElementById("resetGameBtn") as HTMLButtonElement;
+  resetGameBtn.addEventListener("click", () => {
+	  socket.send(JSON.stringify({ type: "resetGame" }));
+  });
+
