--- src/views/pong.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8" />
-	<title>Pong Game</title>
-	<style>
-		body { background: black; margin: 0; overflow: hidden; }
-		canvas { display: block; margin: auto; background: #111; }
-	</style>
-</head>
-<body>
-	<button id="startGameBtn">Start Game</button>
-	<button id="pauseGameBtn">Pause Game</button>
-	<button id="resumeGameBtn">Resume Game</button>
-	<button id="resetGameBtn">Reset Game</button>
-	<canvas id="gameCanvas" width="800" height="600">
-		Your browser does not support the HTML5 Canvas element.
-	</canvas>
-	<script type="module" src="/dist/views/pongMain.js"></script>
-</body>
-</html>
