+++ tools/old/tic_tac_toe.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>ft_transcendence</title>
+    <link rel="icon" href="/public/favicon.ico">
+    <link rel="stylesheet" href="/public/styles.css">
+</head>
+
+<body>
+    <h1>Tic Tac Toe with Typescript</h1>
+    <div class="gameBoard"></div>
+    <button class="restartButton">Restart Game</button>
+    <!-- <script type="module" src="/dist/views/main.js"></script> -->
+    <script type="module" src="/dist/views/main_api.js"></script>
+</body>
+
+</html>
